+++ projects/nested_transactionable/lib/foobara/nested_transactionable.rb
@@ -19,30 +19,13 @@ module Foobara
       end
 
       def with_needed_transactions_for_type(type, &)
-        relevant_entity_classes = relevant_entity_classes_for_type(type)
+        entity_classes = relevant_entity_classes_for_type(type)
 
-        if relevant_entity_classes.empty?
+        if entity_classes.empty?
           return yield
         end
 
-        tx_class = Class.new
-        tx_class.include NestedTransactionable
-
-        tx_class.define_method(:relevant_entity_classes) do
-          relevant_entity_classes
-        end
-
-        tx_instance = tx_class.new
-
-        begin
-          tx_instance.open_transaction
-          result = Persistence::EntityBase.using_transactions(tx_instance.transactions, &)
-          tx_instance.commit_transaction
-          result
-        rescue
-          tx_instance.rollback_transaction
-          raise
-        end
+        TransactionGroup.run(entity_classes:, &)
       end
     end
 
@@ -62,13 +45,26 @@ module Foobara
 
     def opened_transactions
       @opened_transactions ||= []
+      Persistence.sort_transactions(@opened_transactions)
     end
 
     def auto_detect_current_transactions
-      classes = relevant_entity_classes
-      return if classes.nil? || classes.empty?
+      bases = nil
 
-      bases = classes.map(&:entity_base).uniq
+      if respond_to?(:relevant_entity_bases)
+        bases = relevant_entity_bases
+
+        unless bases.nil?
+          return if bases.empty?
+        end
+      end
+
+      unless bases
+        classes = relevant_entity_classes
+        return if classes.nil? || classes.empty?
+
+        bases = classes.map(&:entity_base).uniq
+      end
 
       bases.each do |base|
         tx = base.current_transaction
@@ -84,12 +80,35 @@ module Foobara
 
       bases_not_needing_transaction = transactions.map(&:entity_base)
 
-      bases_needing_transaction = relevant_entity_classes.map(&:entity_base).uniq - bases_not_needing_transaction
+      bases_needing_transaction = nil
+
+      if respond_to?(:relevant_entity_bases)
+        bases_needing_transaction = relevant_entity_bases
+
+        unless bases_needing_transaction.nil?
+          return if bases_needing_transaction.empty?
+        end
+      end
+
+      unless bases_needing_transaction
+        classes = relevant_entity_classes
+        return if classes.nil? || classes.empty?
+
+        bases_needing_transaction = relevant_entity_classes.map(&:entity_base).uniq - bases_not_needing_transaction
+      end
+
+      bases_needing_transaction = Persistence.sort_bases(bases_needing_transaction)
 
       bases_needing_transaction.each do |entity_base|
-        transaction = entity_base.transaction
-        transaction.open!
-        opened_transactions << transaction
+        transaction_mode = if respond_to?(:transaction_mode)
+                             self.transaction_mode
+                           end
+        transaction = entity_base.transaction(transaction_mode)
+        unless transaction.currently_open?
+          transaction.open!
+          @opened_transactions ||= []
+          @opened_transactions << transaction
+        end
         transactions << transaction
       end
     end
@@ -120,6 +139,38 @@ module Foobara
 
     def use_transaction(&)
       Persistence::EntityBase.using_transactions(transactions, &)
+    end
+  end
+
+  class TransactionGroup
+    include NestedTransactionable
+
+    class << self
+      def run(mode: nil, entity_classes: nil, bases: nil, &)
+        new(
+          transaction_mode: mode,
+          relevant_entity_classes: entity_classes,
+          relevant_entity_bases: bases
+        ).run(&)
+      end
+    end
+
+    attr_accessor :relevant_entity_classes, :relevant_entity_bases, :transaction_mode
+
+    def initialize(transaction_mode: nil, relevant_entity_classes: nil, relevant_entity_bases: nil)
+      self.relevant_entity_classes = relevant_entity_classes
+      self.relevant_entity_bases = relevant_entity_bases
+      self.transaction_mode = transaction_mode
+    end
+
+    def run(&)
+      open_transaction
+      result = Persistence::EntityBase.using_transactions(transactions, &)
+      commit_transaction
+      result
+    rescue
+      rollback_transaction
+      raise
     end
   end
 end