+++ projects/persistence/src/entity_base/table.rb
@@ -1,6 +1,7 @@
 module Foobara
   module Persistence
     class EntityBase
+      # TODO: this class seems useless and confusing since we have two other Table classes
       class Table
         attr_accessor :table_name, :entity_base